ML-Python/Server/requirements.txt
flask
pillow
easyocr
opencv-python
numpy
